fix : fixes inconsistencies between the employee portal and admin portal where registration was added by mistake and the right registration component finally replaced the faulty one in the employee portal
