+++ frontend/src/components/auth/EmployeeLoginPage.tsx
@@ -1,4 +1,3 @@
-// src/components/auth/EmployeeLoginPage.tsx
 import React, { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { useNavigate, Link } from 'react-router-dom';
@@ -17,6 +16,8 @@ import {
 import PersonIcon from '@mui/icons-material/Person';
 import { LoginCredentials } from '../../types/auth.types';
 import { useAuth } from '../../context/AuthContext';
+import { api } from '../../services/api';
+import { toast } from 'react-toastify';
 
 interface TabPanelProps {
   children?: React.ReactNode;
@@ -39,13 +40,19 @@ function TabPanel(props: TabPanelProps) {
   );
 }
 
+interface RegisterData extends LoginCredentials {
+  username: string;
+  confirmPassword: string;
+}
+
 export const EmployeeLoginPage: React.FC = () => {
   const [tabValue, setTabValue] = useState(0);
   const { register: loginRegister, handleSubmit: handleLoginSubmit, formState: { errors: loginErrors } } = useForm<LoginCredentials>();
-  const { register: registerRegister, handleSubmit: handleRegisterSubmit, formState: { errors: registerErrors } } = useForm<LoginCredentials>();
+  const { register: registerRegister, handleSubmit: handleRegisterSubmit, formState: { errors: registerErrors }, watch } = useForm<RegisterData>();
   const { login, error, loading } = useAuth();
   const navigate = useNavigate();
   const theme = useTheme();
+  const password = watch('password', '');
 
   const handleTabChange = (event: React.SyntheticEvent, newValue: number) => {
     setTabValue(newValue);
@@ -60,12 +67,26 @@ export const EmployeeLoginPage: React.FC = () => {
     }
   };
 
-  const onRegister = async (data: LoginCredentials) => {
+  const onRegister = async (data: RegisterData) => {
     try {
-      // TODO: Implement registration logic
-      console.log('Registration data:', data);
-    } catch (err) {
-      console.error('Registration error:', err);
+      if (data.password !== data.confirmPassword) {
+        toast.error('Passwords do not match');
+        return;
+      }
+      
+      const response = await api.post('/auth/register/', {
+        email: data.email,
+        username: data.username,
+        password: data.password,
+        role: 'EMPLOYEE',
+        status: 'PENDING' // New users start with pending status
+      });
+      
+      toast.success('Registration successful! Please wait for admin approval to login.');
+      setTabValue(0); // Switch to login tab
+    } catch (err: any) {
+      const errorMessage = err.response?.data?.detail || 'Invalid Operation, Email already exists, or may be some other error';
+      toast.error(errorMessage);
     }
   };
 
@@ -172,9 +193,6 @@ export const EmployeeLoginPage: React.FC = () => {
                 sx={{
                   '& .MuiOutlinedInput-root': {
                     backgroundColor: 'rgba(255, 255, 255, 0.9)',
-                    '&:hover fieldset': {
-                      borderColor: theme.palette.primary.main,
-                    },
                   },
                 }}
               />
@@ -198,31 +216,24 @@ export const EmployeeLoginPage: React.FC = () => {
                 sx={{
                   '& .MuiOutlinedInput-root': {
                     backgroundColor: 'rgba(255, 255, 255, 0.9)',
-                    '&:hover fieldset': {
-                      borderColor: theme.palette.primary.main,
-                    },
                   },
                 }}
               />
+
               <Button
                 type="submit"
                 fullWidth
                 variant="contained"
+                disabled={loading}
                 sx={{
-                  mt: 4,
+                  mt: 3,
                   mb: 2,
                   py: 1.5,
-                  fontSize: '1.1rem',
+                  fontSize: '1rem',
                   fontWeight: 600,
                   textTransform: 'none',
-                  borderRadius: 2,
-                  backgroundColor: theme.palette.primary.main,
-                  '&:hover': {
-                    backgroundColor: theme.palette.primary.dark,
-                  },
-                  boxShadow: '0 4px 20px rgba(0, 0, 0, 0.1)',
+                  boxShadow: '0 4px 12px rgba(0, 0, 0, 0.15)',
                 }}
-                disabled={loading}
               >
                 {loading ? 'Signing in...' : 'Sign In'}
               </Button>
@@ -250,12 +261,28 @@ export const EmployeeLoginPage: React.FC = () => {
                 sx={{
                   '& .MuiOutlinedInput-root': {
                     backgroundColor: 'rgba(255, 255, 255, 0.9)',
-                    '&:hover fieldset': {
-                      borderColor: theme.palette.primary.main,
-                    },
                   },
                 }}
               />
+
+              <TextField
+                margin="normal"
+                required
+                fullWidth
+                id="username"
+                label="Username"
+                {...registerRegister('username', {
+                  required: 'Username is required',
+                })}
+                error={!!registerErrors.username}
+                helperText={registerErrors.username?.message}
+                sx={{
+                  '& .MuiOutlinedInput-root': {
+                    backgroundColor: 'rgba(255, 255, 255, 0.9)',
+                  },
+                }}
+              />
+
               <TextField
                 margin="normal"
                 required
@@ -275,29 +302,42 @@ export const EmployeeLoginPage: React.FC = () => {
                 sx={{
                   '& .MuiOutlinedInput-root': {
                     backgroundColor: 'rgba(255, 255, 255, 0.9)',
-                    '&:hover fieldset': {
-                      borderColor: theme.palette.primary.main,
-                    },
                   },
                 }}
               />
+
+              <TextField
+                margin="normal"
+                required
+                fullWidth
+                label="Confirm Password"
+                type="password"
+                id="confirm-password"
+                {...registerRegister('confirmPassword', {
+                  required: 'Please confirm your password',
+                  validate: value => value === password || 'Passwords do not match'
+                })}
+                error={!!registerErrors.confirmPassword}
+                helperText={registerErrors.confirmPassword?.message}
+                sx={{
+                  '& .MuiOutlinedInput-root': {
+                    backgroundColor: 'rgba(255, 255, 255, 0.9)',
+                  },
+                }}
+              />
+
               <Button
                 type="submit"
                 fullWidth
                 variant="contained"
                 sx={{
-                  mt: 4,
+                  mt: 3,
                   mb: 2,
                   py: 1.5,
-                  fontSize: '1.1rem',
+                  fontSize: '1rem',
                   fontWeight: 600,
                   textTransform: 'none',
-                  borderRadius: 2,
-                  backgroundColor: theme.palette.primary.main,
-                  '&:hover': {
-                    backgroundColor: theme.palette.primary.dark,
-                  },
-                  boxShadow: '0 4px 20px rgba(0, 0, 0, 0.1)',
+                  boxShadow: '0 4px 12px rgba(0, 0, 0, 0.15)',
                 }}
               >
                 Register
@@ -307,17 +347,22 @@ export const EmployeeLoginPage: React.FC = () => {
 
           <Button
             component={Link}
-            to="/admin"
+            to="/"
+            variant="outlined"
+            fullWidth
             sx={{
               mt: 2,
-              color: theme.palette.text.secondary,
-              textDecoration: 'none',
+              py: 1.5,
+              textTransform: 'none',
+              borderColor: theme.palette.primary.main,
+              color: theme.palette.primary.main,
               '&:hover': {
-                color: theme.palette.primary.main,
+                borderColor: theme.palette.primary.dark,
+                backgroundColor: 'rgba(0, 0, 0, 0.04)',
               },
             }}
           >
-            Admin Login
+            Admin Portal
           </Button>
         </Paper>
       </Container>
